backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
django==5.1.3
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   djangorestframework
django-cors-headers==4.6.0
    # via -r requirements.in
django-environ==0.11.2
    # via -r requirements.in
djangorestframework==3.15.2
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
packaging==24.1
    # via gunicorn
pillow==11.0.0
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
sqlparse==0.5.1
    # via django
tzdata==2024.2
    # via django
whitenoise==6.8.2
    # via -r requirements.in
